requirements.txt
streamlit
pandas
numpy
streamlit
google-generativeai
pdfplumber
gspread
google-auth